pyproject.toml
[project]
name = "mcpdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "mcp-use>=1.7.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.1",
]
